+++ Modules/Registration/Metricsv4/test/itkMeanSquaresImageToImageMetricv4RegistrationTest.cxx
@@ -61,7 +61,7 @@ int itkMeanSquaresImageToImageMetricv4RegistrationTest(int argc, char *argv[])
     }
   if( argc >= 6 )
     {
-    numberOfDisplacementIterations = atof( argv[5] );
+    numberOfDisplacementIterations = atoi( argv[5] );
     }
   std::cout << " iterations "<< numberOfIterations
     << " displacementIterations " << numberOfDisplacementIterations << std::endl;
